requirements.txt
# Dependencias principales
flask==2.3.3
flask-sqlalchemy==3.1.1
reportlab==4.0.7
google-api-python-client==2.107.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pydrive2==1.17.0
setuptools==68.2.2

# Dependencias para serverless en Netlify
netlify-lambda==0.4.0
serverless-wsgi==3.0.2

# Otras utilidades
python-dotenv==1.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
SQLAlchemy==2.0.23